fixing merge conflicts to merge master into current
