--- app/namespaces/airtable_scraper/airtable_scraper_schema.py
@@ -1,6 +1,0 @@
-from marshmallow import Schema, fields
-
-
-class StudyCountSchema(Schema):
-    records = fields.List(fields.Dict, required=True)
-
