+++ app/namespaces/data_provider/data_provider_schema.py
@@ -21,3 +21,7 @@ class RecordsSchema(Schema):
 class RecordDetailsSchema(Schema):
     source_id = fields.UUID(required=True)
 
+
+class StudyCountSchema(Schema):
+    records = fields.List(fields.Dict, required=True)
+
